requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
pandas==2.1.3
numpy==1.26.2
websockets==12.0 